requirements.txt
undetected-chromedriver
selenium
beautifulsoup4
pyyaml
